Fix error when hitting the tasks endpoint with no parameters
There were some inconsistencies in the custom REST Controller class
that caused the `sanitize_post_status` method to always return an error
if the endpoint was hit from an unauthenticated session.

Now, the endpoint still returns an empty set for an unauthenticated
session, which may be due to the custom statuses used by tasks. But
at least it's not an error. This isn't much of a problem because this
API endpoint is only intended for use while logged into WP Admin anyway.