Modify REST routes and permissions checks
Before we were relying on the routes and permissions in the Posts
controller. However, since we have specific capabilities for
Organizers and Mentors, and we don't need to create or delete tasks
via the REST API, this will simplify the endpoints and also prevent
public, unauthenticated access to the task data.